Patch from Jason Diamond to make win32 platform stuff work:
* Modified Configure.pl, mswin32.pl, and Makefile.in so that the platform
specific files in the platforms directory gets copied to the correct
directories. The Makefile will re-copy the platform.h and .c files if the
original is modified.

* #include platform.h in parrot.h instead of time.h.

* Removed time.h and time.obj from the Makefile. time.h and time.c aren't
being used anymore since that code was moved to linux.c and win32.c.

* Renamed the BOOL typedef in parrot.h to BOOLVAL (inspired by INTVAL and
FLOATVAL). This was conflicting with the BOOL typedef in windows.h. This
required updating several files besides parrot.h.

* Added a DEFAULT_OPEN_MODE to linux.h and win32.h and used that instead of
the missing identifiers (on Windows) in core.ops in the calls to open.

* Added Parrot_sleep and Parrot_setenv to both the linux and win32 platform
files and called those from core.ops. The Linux code for these functions
came from core.ops so they should hopefully work.

Courtesy of Jason Diamond <jason@injektilo.org>


git-svn-id: https://svn.parrot.org/parrot/trunk@389 d31e2699-5ff4-0310-a27c-f18f2fbe73fe