This patch adds a few more macros for .ops files to use. In addition to the existing
 goto OFFSET(...)
 goto ADDRESS(...)
 goto POP()
 goto NEXT()

it adds

 expr OFFSET(...)
 expr ADDRESS(...)
 expr POP()
 expr NEXT()

This is intended mainly for use in helper functions in .ops files, but
also for ops that need to eg push the address of the next opcode onto
a stack. The only place this currently happens (outside of my private
code) is in the bsr opcode, which previously hardcoded the
computation.

This patch also makes ops2c.pl move the CUR_OPCODE #define a bit
higher in the generated file so that it can be used by static
functions in the preamble.

Finally, I ran into a parentheses matching problem -- should goto
OFFSET(...) rewrite
   \bgoto\s+OFFSET\((.*)\)
or
   \bgoto\s+OFFSET\((.*?)\)

It was done inconsistently for the different calculations, and of
course neither are correct because you really want matching
parentheses. But that seems like too much bother. (Currently, there's
only one place that uses anything other than $n in the parentheses,
and that's the enternative() op.) So I made them all .*? but emulated
recent POD by also allowing (( ... )) (that's two parenthesis with a
space padding the body.)

The generated core_ops.c and core_ops_prederef.c files after this
patch are identical except for the CUR_OPCODE and REL_PC macros
getting moved a little higher, and the bsr in the regular path (not
prederef) uses 'CUR_OPCODE' instead of the hardcoded 'cur_opcode'.

In summary, this is a big wad of meaningless changes for the current
code. But it will make adding some opcodes easier in the future, and
cleans up a small nit in the current bsr.

Courtesty of Steve Fink <steve@fink.com>


git-svn-id: https://svn.parrot.org/parrot/trunk@846 d31e2699-5ff4-0310-a27c-f18f2fbe73fe