Attached is a simplified version of a previous patch to allow buffers to avoid collection during their formative nanoseconds. This version covers the basics: A new flag - BUFFER_neonate_FLAG This flag causes immunity from collection during DOD runs The flag is set and cleared as required in string.c (note that there may be more places that need flag setting, I just did a few obvious ones)
This version makes no attempt to track the fact that there are newborns, or
to kill them if they try to keep this status for too long. It therefore
needs to be specifically used only when required. I am proposing this as an
interim solution until we decide on the best way of handling it permanently,
to overcome the current bugs without resorting to the performance-impacting
method of suppressing DOD runs.

A few warnings in string.c have been removed in the process, along with some
duplicate #define's in string.h

Courtesy of: Peter Gibbs <peter@emkel.co.za>


git-svn-id: https://svn.parrot.org/parrot/trunk@1498 d31e2699-5ff4-0310-a27c-f18f2fbe73fe