Many apologies for the size of the assemble.pl diffs, but it's either that or lose the versioning, which I'm loathe to do.
root.in got modified because the new version of assemble.pl requires the XS stuff to be compiled in. The tests were modifed only because of the new syntax for macros and local labels. More features to be added later on, but for now the languages/ directory probably needs some serious munging. I'll do what I can, but the owners of the respective languages should be able to do most of their own work, I hope.

This is a biggie, but despite the number of files affected, I'm hoping it won't be that much in the long run. This should get us a long way towards support for
a lot of the stuff we'll need for 1.0.


git-svn-id: https://svn.parrot.org/parrot/trunk@1559 d31e2699-5ff4-0310-a27c-f18f2fbe73fe