This seems like a good time to send in this patch:
It allocates the stack content memory using a buffer. This makes the stack
chunks and the memory used to hold stack contents visible to the garbage
collector.  One can incrementally add to this to support copy-on-write
semantics for the chunk contents, which I understand is going to be useful in
taking continuations. This could be done by making the stack chunks themselves
buffer headers or perhaps PMCs, which would be cool for introspection.

Courtesy of Jason Gloudon <perl@gloudon.com>


git-svn-id: https://svn.parrot.org/parrot/trunk@1686 d31e2699-5ff4-0310-a27c-f18f2fbe73fe