Two changes:
a) minor doc fix in core.ops:
mem_allocs_since_last_collect is the number of new blocks allocated,
not the total memory allocated.

b) fix for the following problem:
The title says it all really: the counter in the interpreter structure
that tracks recent header allocations is initialized to 0 when the
interpreter is set up, but isn't incremented when headers are allocated.
Consequently, this:

interpinfo I1, 8
print I1

always prints zero.


Courtesy of Simon Glover.


git-svn-id: https://svn.parrot.org/parrot/trunk@2001 d31e2699-5ff4-0310-a27c-f18f2fbe73fe