I'm ba-ack :)
Anyway, here are not one but *two* grammars for Ruby.

The first grammar, located in the regular directory, is a grammar reverse-engineered (due to potential license conflicts) from the available documentation. Neither grammar has had all of the nasty left-recursiveness removed from it, but the first grammar is the closest, as only 10 lines in the entire grammar are left-recursive. Using the 'interactive.pl' tool (after installing YAML or switching the appropriate lines to use Data::Dumper), the simple statement 'a=1' descends about five levels into the grammar.

The other grammar is in the Yacc/ directory. This is taken from the official parse.y grammar file, stripped down, and given a rather exhaustive test suite. The 01_terminal.t and 02_expression.t files are pretty extensive, but nowhere near ready for prime time.

For the purposes of development, I think the stuff in the Yacc/ directory, derived as it is directly from the parse.y file, is the ultimate way to go, but I encourage you to wait for approval before hacking on it.


git-svn-id: https://svn.parrot.org/parrot/trunk@2076 d31e2699-5ff4-0310-a27c-f18f2fbe73fe