In tracking down a gc bug, I realized that the current throwaway implementation of the print op could be replaced with a faster throwaway implementation that avoids doing a string_to_cstring.
Note that both the original and new implementations are still buggy
with respect to supporting different encodings. I don't know if
printf("%s") is any better than fwrite in terms of at least vaguely
paying attention to your locale or whatever. If so, don't apply it.

(Courtesy of Steve Fink)


git-svn-id: https://svn.parrot.org/parrot/trunk@2268 d31e2699-5ff4-0310-a27c-f18f2fbe73fe