This update includes three patches:
Ticket 16858:
Fixes for resources.c to not copy the buffer if it is
BUFFER_external_FLAG.


Ticket 16859:
This patch is the real fix for strings with the BUFFER_external_FLAG.
It requires the previous dod.c and resources.c patches to be applied.

Strings marked with the BUFFER_external_FLAG point to the external
memory rather than making a copy for themselves. Side effects of this
are (1) less memory usage, (2) external updates to the string are
reflected in the Parrot STRING (except for length changes!), and (3)
these strings are skipped over when memory is getting moved around
during a compaction.


My own magic patch ticket:
It removes the selfpoolptr_FLAG and its cool comment, in favor of
BUFFER_external_FLAG.


All tests pass.

Patches one and two courtesy of Steve Fink.


git-svn-id: https://svn.parrot.org/parrot/trunk@2270 d31e2699-5ff4-0310-a27c-f18f2fbe73fe