This patch get's rid of the +1000 lines static hash in core_ops.c and in core_ops_prederef.c.
On first invocation of op_lib->op_code(), a dynamic hash of fullname's
_and_ short name's gets built, which is used for lookup.

There is a slight change in the interface, the second parameter tells to
search for a "fullname" or a "name". This is useful (and needed for
imcc), to allow the lexer, to discern an indentifier from an op_code.

Courtesy of Leopold Toetsch <lt@toetsch.at>


git-svn-id: https://svn.parrot.org/parrot/trunk@2306 d31e2699-5ff4-0310-a27c-f18f2fbe73fe