Leopold Toetsch's big PMC reorganization. Comment is a summary of various things Leo said about the pieces of this patch, plus a summary and a description of what I modified before committing this:
Changes the PMC enumeration
so that:
- default.pmc is implicitly #0
- it is illegal to "new P0, 0"
- it is illegal to "new P0, $enum_class_max"

This is the base for classes/*.pmc clean up,
s. also
"[RFC] How are compound keys with a PerlHash intended to work?"

Tests for corner cases are included.

Makes almost all methods in default.pmc spit out an exception. The
famous "Subscript on something, that is not an aggregate" now looks
like, e.g.: "set_integer_keyed() not implemented in class 'PerlInt'"

- cleans up array and perlarray
- makes consistent resizing in perlarray
- includes Sean's patch to perlarray
- reuses more array methods in perlarray

My (sfink's) changes:
 - Fixed a few typos
 - Implemented modulus for perlnum (for no good reason)
 - %p used in a format string in place of 0x%x
 - Removed some duplicate tests
 - The only significant thing: the configuration system previously
   could not handle a nontrivial inheritance hierarchy, so things
   could get built out of order and therefore fail to
   compile. config/inter/pmc.pl now traces out the inheritance
   hierarchy to add the appropriate dependencies to the
   classes/Makefile.

The current class hierarchy now looks like:

default
    scalar
        perlint
            boolean
            perlundef
        perlnum
        perlstring
    array
        perlarray
    continuation
    coroutine
    csub
    intlist
    intqueue
    key
    multiarray
    perlhash
    pointer
    sub

I believe the short-term plan is to replace intlist with list and make
it the parent of array and perlarray.


git-svn-id: https://svn.parrot.org/parrot/trunk@2394 d31e2699-5ff4-0310-a27c-f18f2fbe73fe