Trying to run parrot with the PARROT_GC_DEBUG environment variable set to 1 causes parrot to fail during initialization.
The GC_DEBUG flag is set early based on the environment variable, but the
marker for the top of the C stack is not set so non-rooted PMCs created during
initialization are wrongly collected.

The patch sets the stack top pointer at the time the GC_DEBUG flag is set.

Courtesy of Jason Gloudon <perl@gloudon.com>`


git-svn-id: https://svn.parrot.org/parrot/trunk@2863 d31e2699-5ff4-0310-a27c-f18f2fbe73fe