Parrot BASIC now supports the cls command to clear the screen.  It works under Win32 by using NCI to access kernel32.dll to get to the Console manipulation commands.  (Inkey$, Locate, csrline, pos, beep, color and others will follow shortly)
For the moment I can't tell what platform the PASM is running on and I can't trap
loadlib (it crashes the VM).  If you want to look around and you're not on Win32 you
need to avoid these functions for now (cls, etc..).  The rest of BASIC should be fine.


git-svn-id: https://svn.parrot.org/parrot/trunk@3228 d31e2699-5ff4-0310-a27c-f18f2fbe73fe