This is the first step of the IO-system away from the mem_sys_alloc/free memory-managment system to a full-fledged PMC-based system.
In this patch only the ParrotIO structures are transformed to a
PMC. This is simply done by wrapping the ParrotIO in a PMC_data. For
easier access of the layer-pointer (this is the first thing that is
actually needed if a function is called) is stored in the
cache.struct_val. The rest of the patch is mostly to an update of
parrot to the new prototypes.

The filedescriptor-ops are not deleted in this patch, but use
a hack to always create a new temporary PMC from the filedescriptor.

The standard filedescriptors stdin, stdout and stderr are protected
from closing by the PMC_destroy function by marking them as
PIO_F_SHARED. Don't know if this is a right solution.

Courtesy of J��rgen B��mmels <boemmels@physik.uni-kl.de>

(Committer's note: I think this exposes some latent bugs, as it fails
9 tests on OS X, but getting it in is worth it so other folks can
thump the thing)


git-svn-id: https://svn.parrot.org/parrot/trunk@3496 d31e2699-5ff4-0310-a27c-f18f2fbe73fe