Makefile cleanup:
languages.in:
- add $(MAKE_C) befunge clean in target 'clean'

root.in:
- add target 'help': This should give a basic overview of what can be done
with the Makefile
- add target 'languages-clean': I needed that for testing
- add target 'docs-clean': This seems to be consistent
- add a '@' in 'flags-dummy': "Compiling with" should be printed only once
- add target 'patch': generate a patch, when there are diffs to CVS, seems
to be useful to me

I also did some reformating of lists in 'root.in'. This is of course a
matter of taste.
I don't like tabs in lists. As tabs have a defined meaning in targets, I
find it confusing to see them in macro assignments.

Courtesy of "Bernhard Schmalhofer" <Bernhard.Schmalhofer@gmx.de>


git-svn-id: https://svn.parrot.org/parrot/trunk@4916 d31e2699-5ff4-0310-a27c-f18f2fbe73fe