#24823
GNU m4, http://www.gnu.org/software/m4, is a macro processor used for
example in GNU autoconf.
Parrot m4 is a direct port of GNU m4. Thus Parrot m4 is GPL licensed, which
should be OK for stuff in
'parrot/languages'. The implementation language is PIR.

The current status is very much pre-release. There is a very long TODO list
and scant documentation.
The last major change is the switch from line-based input reading to a more
proper parsing with 'libpcre'.
So 'libpcre' needs to be installed for running Parrot m4.
For running the tests, GNU m4 needs to available in $PATH.

Courtesy of Bernhard Schmalhofer


git-svn-id: https://svn.parrot.org/parrot/trunk@5042 d31e2699-5ff4-0310-a27c-f18f2fbe73fe