[perl #28454] [PATCH] Example 'japh16.pasm' broken after integration of ICU
Editing linux.pl and defining 'c++' as the linker works for me. Editing the
Makefile and adding '-lstdc++' worked also.

It looks like the signature of 'string_make' has changed. This broke the the
loadable compiler 'japhc.so' used for japh16.pasm.
A patch fixing that problem is attached.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>

And fixed 2 more broken JAPHs (leo)


git-svn-id: https://svn.parrot.org/parrot/trunk@5859 d31e2699-5ff4-0310-a27c-f18f2fbe73fe