[perl #29502] [PATCH] expanding arguments of macro's
The first bug has to do with nested macro calls that use the same name
as parameter.
Expansion of .X results in an endless loop.

The second bug has to do with a local label as argument of a macro call.

The patch provided here as attachment solves both bugs.
It puts some code in the function expand_macro that expands the arguments
of the macro.

Courtesy of Bob Diertens <bobd@science.uva.nl>


git-svn-id: https://svn.parrot.org/parrot/trunk@6023 d31e2699-5ff4-0310-a27c-f18f2fbe73fe