[perl #30534] [PATCH] Complex numbers PMC [perl #30535] [PATCH] Complex numbers PMC
This is the new classes/complex.pmc, tests included. It supports string
parsing (in the form "a + bi", see the docs for more), string-keyed
access to "real" and "imag" and simple numerical operations. However,
the function get_number (it should return the modulus) is not
implemented, because I couldn't find sqrt() anywhere. Is it OK if I use
the one in the standard library?

[ patch was attached at the 2nd ticket - both resolve ]

Courtesy of Ion Alexandru Morega <alexm@iem.pub.ro>


git-svn-id: https://svn.parrot.org/parrot/trunk@6293 d31e2699-5ff4-0310-a27c-f18f2fbe73fe