[perl #33751] [PATCH] Use INTERP in *.pmc files
I noticed the there is an interesting mix of 'interpreter' and 'INTERP' in
the vtable-functions in the *.pmc files. This works, because
'classes/pmc2c2.pl' changes
'INTERP' to 'interpreter'.
For the sake of consistency I have changed 'interpreter' to INTERP.

Also, the 'Parrot_Interp' declarations in the *.pmc files are changed to
'Interp*'. I think that this is the recommended declaration for use in the
Parrot core.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>

[ applied sans dynclassess ]


git-svn-id: https://svn.parrot.org/parrot/trunk@7455 d31e2699-5ff4-0310-a27c-f18f2fbe73fe