[perl #34051] [PATCH] return absolute paths from Parrot_locate_runtime_file
Parrot_readbc can segfault when loading an absolute path to a file (in my
case, only under GDB).  the problem lies in Parrot_locate_runtime_file,
and is similar to the one in ticket #32087.

the attached patch adds a check in Parrot_locate_runtime_file so it
returns absolute paths as is.  it handles win32 paths (and drive letters)
as well.  hopefully i chose the most logical place for this -- i didn't
think it belonged in Parrot_readbc itself.

Courtesy of Jeff Horwitz <jeff@smashing.org>


git-svn-id: https://svn.parrot.org/parrot/trunk@7530 d31e2699-5ff4-0310-a27c-f18f2fbe73fe