[perl #34117] [PATCH] fix make html
make html failes currently, because the src/test_main.c file was moved
to examples/c/text_main.c.

ChangeLog:
   - update search path to find the file test_main.c.
   - warn if a file is not found


[perl #34120] [PATCH] win32 bind, listen, accept

This patch extends the io NET_DEVEL with the 'server' functions bind,
listen and accept. I only tested it on win32, so no idea what happens on
unix. I'm not sure wether i did it right with the ops numbering, but it
worked for me (See the next patch).

Changelog:
  - implement bind, listen, accept on win32
  - some casting cleanup


[perl #34121] [NEW] imc http server

Now it's getting funny. I have written a tiny webserver in imc, that can
serve the parrot html documentation. That's also a pretty good test for
the NET_DEVEL and file readings functions. I tested it on win32, so no
idea what happens on unix.
The attached file should be placed in examples/io (no idea how to create
a /dev/null patch on win32!).

This patch depends of course on #34120.

[perl #34126] [PATCH] fix unix bind, listen, accept

With this patch the tiny webserver from # 34121 runs also on linux
(tested on debian testing).

depends on #34120

Changelog:
  - fix bind, listen, accept on unix
  - remove some debug output

All
Courtesy of Markus Amslser <markus.amsler@oribi.org>


git-svn-id: https://svn.parrot.org/parrot/trunk@7532 d31e2699-5ff4-0310-a27c-f18f2fbe73fe