2 patches
[PATCH]Parrot_dod_sweep fix

This patch prevents from an attempt to add a NULL PMC_EXT structure to the
pmc ext pool while in dod sweep run.

[PATCH]PackFile_destroy fix

Nested packfile segments of type directory are freed twice by
default_destroy(). The first time it happens in directory_destroy(), and the
second time in PackFile_Segment_destroy(). This behavior is, of course,
incorrect, the

Courtesy of Vladimir Lipsky <searcher768@rambler.ru>



git-svn-id: https://svn.parrot.org/parrot/trunk@8210 d31e2699-5ff4-0310-a27c-f18f2fbe73fe