[PATCH] Win32 executable generation
The attached patch does some changes to make it possible to generate an EXE
file on Win32.

Firstly, main.c looked for the extension .o to know to make an object file;
on Win32 it's .obj.  The fix will work for other platforms where it isn't .o
too.

Secondly, the make exec target in the makefile didn't append the .exe to the
generated file.

Courtesy of Jonathan Worthington <jonathan@jwcs.net>


git-svn-id: https://svn.parrot.org/parrot/trunk@8672 d31e2699-5ff4-0310-a27c-f18f2fbe73fe