tcl: #38068: [TODO] Tcl: Add support for octal numbers
o Add support for octal numbers wherever an integer is expected.
  (notably [expr], [incr]), plus tests.

Also:

o Add tests for unary ops in [expr], fix a bug with [expr -2]
o rename a .sym in a macro to avoid collision with shorter varname.
o fix bug with [cd] that crept in yesterday in the makefile



git-svn-id: https://svn.parrot.org/parrot/trunk@10919 d31e2699-5ff4-0310-a27c-f18f2fbe73fe