[GC] Extracted several PObj-freeing functions from Parrot_dod_sweep().
The next step is to regularlize their APIs (mostly by passing in the pool), and
then making them function pointers attached to the pool structures.  This
should remove both the need to perform "Hey, what type is this?" checks for
every PObj in a pool, as well as the possibility of performing the wrong action
on the wrong type of PObj.

git-svn-id: https://svn.parrot.org/parrot/trunk@20808 d31e2699-5ff4-0310-a27c-f18f2fbe73fe