[IMCC] Plugged most of the larger leaks in IMCC.  It leaks about half as much memory as when Andy and I started today.
These mem_sys_free() calls are mostly bandages on the real problem, which is
the str_dup() call for the STRINGCONSTANT lexing rule in imcc.l.  Something in
the parser (imcc.y) passes this value unmodified to a function that stores it
directly.  If we can find what that is and do the copy there, we can remove all
of the str_dup() and mem_sys_free() calls scattered around the handling of
STRINGC, USTRINGC, and IDENTIFIER tokens in the parser.

This means you, reading this, right now.  Hop to it!

git-svn-id: https://svn.parrot.org/parrot/trunk@25525 d31e2699-5ff4-0310-a27c-f18f2fbe73fe