[src] Fixed clone op not to share PMC_EXT structure between clone source and its ultimate destination.  When the source PMC gets freed (during the next GC run), the GC will free its PMC_EXT structure.  Sadly, because both structures share the pointer, this means that it would be invalid in the clone destination, which is the PMC where you most want it active.
The solution is to clear that pointer in the clone source before any GC runs
can happen.  This fixes the big problem in RT #51136.

I also added yet another destroy call so as to leak even less memory in clone.

There's probably still a problem with the default morph() vtable entry for PMCs
such as Ref and Object, but one thing at a time.

git-svn-id: https://svn.parrot.org/parrot/trunk@26197 d31e2699-5ff4-0310-a27c-f18f2fbe73fe