make toosl/build/dynpmc.pl a static file instead of dynamically generating it.
The original version probably predates Parrot::Config.

While doing this, simplify the logic a bit. (" is always safe to use on
our core platforms. use $^X instead of config's perl, since we're invoked
with the right perl anyway.)

There is no doubt more simplifications to be had here, as we seem to be
calculating an awful lot. 



git-svn-id: https://svn.parrot.org/parrot/branches/one_make@43345 d31e2699-5ff4-0310-a27c-f18f2fbe73fe